# api/openapi.yaml
openapi: 3.0.3
info:
  title: MicroPlay Hub API
  description: API for the MicroPlay Hub social gaming platform
  version: 1.0.0
  contact:
    name: MicroPlay Hub Team
    url: https://github.com/kawacukennedy/microplay-hub
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.microplay-hub.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /moderation/levels:
    get:
      summary: Get pending levels for moderation
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Pending levels
          content:
            application/json:
              schema:
                type: object
                properties:
                  levels:
                    type: array
                    items:
                      $ref: '#/components/schemas/Level'
                  total:
                    type: integer
  /moderation/levels/{id}/approve:
    post:
      summary: Approve level
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                moderatorId:
                  type: string
                notes:
                  type: string
              required:
                - moderatorId
      responses:
        '200':
          description: Level approved
  /moderation/levels/{id}/reject:
    post:
      summary: Reject level
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                moderatorId:
                  type: string
                reason:
                  type: string
                notes:
                  type: string
              required:
                - moderatorId
                - reason
      responses:
        '200':
          description: Level rejected
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [USER, CREATOR, MODERATOR, ADMIN]
        isGuest:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
...
    Level:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        creatorId:
          type: string
        title:
          type: string
        description:
          type: string
        data:
          type: object
        thumbnailUrl:
          type: string
        assets:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        publishedAt:
          type: string
          format: date-time
        isFlagged:
          type: boolean
        playsCount:
          type: integer
        likesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        creator:
          $ref: '#/components/schemas/User'
...